+++ addOns/llm/CHANGELOG.md
@@ -10,4 +10,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support for Google Gemini and OpenRouter.
 - Integration points for other add-ons.
 - Support for logging all LLM comms to a sub-tab of the main Output tab.
-- An LLM Chat panel.
+- A tabbed LLM Chat panel.